move recording timer logic into audio recording service
