+++ RecNForget.Controls/TimeSpanTextBox.xaml.cs
@@ -1,20 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using RecNForget.Services.Contracts;
+using System;
 using System.ComponentModel;
 using System.Globalization;
 using System.Runtime.CompilerServices;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace RecNForget.Controls
 {
@@ -31,7 +22,6 @@ namespace RecNForget.Controls
     /// </summary>
     public partial class TimeSpanTextBox : UserControl, INotifyPropertyChanged
     {
-        public static string ParseFormat = "d':'hh':'mm':'ss";
         public event PropertyChangedEventHandler PropertyChanged;
 
         private void OnPropertyChanged([CallerMemberName] string propertyName = null)
@@ -116,7 +106,7 @@ namespace RecNForget.Controls
             tbEntry.Text = workingValue;
             tbEntry.CaretIndex = tbEntry.Text.Length;
 
-            var parseSuccessful = TimeSpan.TryParseExact(workingValue, ParseFormat, CultureInfo.InvariantCulture, out TimeSpan outputTimeSpan);
+            var parseSuccessful = TimeSpan.TryParseExact(workingValue, Formats.TimeSpanFormat, CultureInfo.InvariantCulture, out TimeSpan outputTimeSpan);
             ValidationErrorMark.Visibility = parseSuccessful ? Visibility.Hidden : Visibility.Visible;
         }
     }
