+++ RecNForget.Services.Contracts/Formats.cs
@@ -1,0 +1,7 @@
+﻿namespace RecNForget.Services.Contracts
+{
+    public class Formats
+    {
+        public const string TimeSpanFormat = "d':'hh':'mm':'ss";
+    }
+}
