+++ RecNForget.Services.Contracts/IAudioRecordingService.cs
@@ -5,6 +5,10 @@ namespace RecNForget.Services.Contracts
 {
     public interface IAudioRecordingService : INotifyPropertyChanged
     {
+        string CurrentRecordingStartAfterTimer { get; set; }
+        string CurrentRecordingStopAfterTimer { get; set; }
+        bool TimerForRecordingStartAfterNotRunning { get; set; }
+        bool TimerForRecordingStopAfterNotRunning { get; set; }
         string CurrentFileName { get; }
 
         string LastFileName { get; }
@@ -21,5 +25,10 @@ namespace RecNForget.Services.Contracts
         void StopRecording();
 
         string GetTargetPathTemplateString();
+        void StartTimerToStartRecordingAfter();
+        void StartTimerToStopRecordingAfter();
+        void ResetStartAfterDispatcherTimer();
+        void ResetStopAfterDispatcherTimer();
+        void ResetAllTimers();
     }
 }
