+++ RecNForget.WPF.Services.Contracts/IActionService.cs
@@ -7,10 +7,6 @@ namespace RecNForget.WPF.Services.Contracts
     public interface IActionService : INotifyPropertyChanged
     {
         System.Windows.Controls.Control OwnerControl { get; set; }
-        string CurrentRecordingStartAfterTimer { get; set; }
-        string CurrentRecordingStopAfterTimer { get; set; }
-        bool TimerForRecordingStartAfterNotRunning { get; set; }
-        bool TimerForRecordingStopAfterNotRunning { get; set; }
 
         // change output directory (with dialog)
         void ChangeOutputFolder();
@@ -59,11 +55,6 @@ namespace RecNForget.WPF.Services.Contracts
         void ShowNewToVersionDialog(Version currentFileVersion, Version lastInstalledVersion);
         void ShowRandomApplicationTip();
         void ExportSelectedFile();
-        void StartTimerToStartRecordingAfter();
-        void StartTimerToStopRecordingAfter();
         void ToggleRecordingTimerControlVisibility();
-        void ResetStartAfterDispatcherTimer();
-        void ResetStopAfterDispatcherTimer();
-        void ResetAllTimers();
     }
 }
