+++ RecNForget.WPF.Services/ActionService.cs
@@ -37,23 +37,6 @@ namespace RecNForget.WPF.Services
         private readonly IApplicationHotkeyService hotkeyService = null;
 
         private readonly NotificationManager _notificationManager = new NotificationManager();
-
-        private DispatcherTimer startAfterDispatcherTimer = new DispatcherTimer();
-        private DispatcherTimer stopAfterdispatcherTimer = new DispatcherTimer();
-
-        private string GetCurrentStartAfterDispatcherTimeString()
-        {
-            return startAfterDispatcherTimerCurrentTime.ToString(TimeSpanTextBox.ParseFormat);
-        }
-
-        private string GetCurrentStopAfterDispatcherTimeString()
-        {
-            return stopAfterDispatcherTimerCurrentTime.ToString(TimeSpanTextBox.ParseFormat);
-        }
-
-        private TimeSpan startAfterDispatcherTimerCurrentTime = TimeSpan.Zero;
-        private TimeSpan stopAfterDispatcherTimerCurrentTime = TimeSpan.Zero;
-
         public Control OwnerControl { get; set; }
 
         private void OnPropertyChanged([CallerMemberName] string propertyName = null)
@@ -72,143 +55,6 @@ namespace RecNForget.WPF.Services
 
             hotkeyService.ResetAndReadHotkeysFromConfig(this);
 
-            startAfterDispatcherTimer.Interval = TimeSpan.FromSeconds(1);
-            startAfterDispatcherTimer.Tick += StartAfter_DispatcherTimer_Tick;
-            stopAfterdispatcherTimer.Interval = TimeSpan.FromSeconds(1);
-            stopAfterdispatcherTimer.Tick += StopAfter_DispatcherTimer_Tick;
-
-        }
-
-        private bool timerForRecordingStartAfterNotRunning = true;
-        public bool TimerForRecordingStartAfterNotRunning
-        {
-            get { return timerForRecordingStartAfterNotRunning; }
-            set { timerForRecordingStartAfterNotRunning = value; OnPropertyChanged(); }
-        }
-
-        private bool timerForRecordingStopAfterNotRunning = true;
-
-        public bool TimerForRecordingStopAfterNotRunning
-        {
-            get { return timerForRecordingStopAfterNotRunning; }
-            set { timerForRecordingStopAfterNotRunning = value; OnPropertyChanged(); }
-        }
-
-
-        private string currentRecordingStopAfterTimer = "0:00:00:00";
-
-        public string CurrentRecordingStopAfterTimer
-        {
-            get { return currentRecordingStopAfterTimer; }
-            set { currentRecordingStopAfterTimer = value; OnPropertyChanged(); }
-        }
-
-        private string currentRecordingStartAfterTimer = "0:00:00:00";
-
-        public string CurrentRecordingStartAfterTimer
-        {
-            get { return currentRecordingStartAfterTimer; }
-            set { currentRecordingStartAfterTimer = value; OnPropertyChanged(); }
-        }
-
-        private void StartAfter_DispatcherTimer_Tick(object sender, EventArgs e)
-        {
-            startAfterDispatcherTimerCurrentTime = startAfterDispatcherTimerCurrentTime.Subtract(TimeSpan.FromSeconds(1));
-            CurrentRecordingStartAfterTimer = GetCurrentStartAfterDispatcherTimeString();
-
-            if (startAfterDispatcherTimerCurrentTime.TotalSeconds < 1)
-            {
-                if (!audioRecordingService.CurrentlyRecording)
-                {
-                    audioRecordingService.ToggleRecording();
-
-                    if (appSettingService.RecordingTimerStopAfterIsEnabled)
-                    {
-                        StartTimerToStopRecordingAfter();
-                        return;
-                    }
-                }
-
-                ResetStartAfterDispatcherTimer();
-            }
-        }
-
-        private void StopAfter_DispatcherTimer_Tick(object sender, EventArgs e)
-        {
-            stopAfterDispatcherTimerCurrentTime = stopAfterDispatcherTimerCurrentTime.Subtract(TimeSpan.FromSeconds(1));
-            CurrentRecordingStopAfterTimer = GetCurrentStopAfterDispatcherTimeString();
-
-            if (stopAfterDispatcherTimerCurrentTime.TotalSeconds < 1)
-            {
-                // ToDo more than two timer options?
-                // timer finished, trigger respective method either way
-                if (!TimerForRecordingStartAfterNotRunning)
-                {
-                    if (!audioRecordingService.CurrentlyRecording)
-                    {
-                        audioRecordingService.ToggleRecording();
-
-                        if (appSettingService.RecordingTimerStopAfterIsEnabled)
-                        {
-                            StartTimerToStopRecordingAfter();
-                            return;
-                        }
-                    }
-                }
-                else
-                {
-                    if (audioRecordingService.CurrentlyRecording)
-                    {
-                        audioRecordingService.ToggleRecording();
-                    }
-                }
-
-                ResetStopAfterDispatcherTimer();
-            }
-        }
-
-        public void StartTimerToStartRecordingAfter()
-        {
-            // reset all other possibly running timers
-            ResetStartAfterDispatcherTimer();
-            startAfterDispatcherTimerCurrentTime = TimeSpan.ParseExact(appSettingService.RecordingTimerStartAfterMax, TimeSpanTextBox.ParseFormat, CultureInfo.InvariantCulture);
-            CurrentRecordingStartAfterTimer = GetCurrentStartAfterDispatcherTimeString();
-            TimerForRecordingStartAfterNotRunning = false;
-
-            startAfterDispatcherTimer.Start();
-        }
-
-        public void StartTimerToStopRecordingAfter()
-        {
-            // reset all other possibly running timers
-            ResetStopAfterDispatcherTimer();
-            stopAfterDispatcherTimerCurrentTime = TimeSpan.ParseExact(appSettingService.RecordingTimerStopAfterMax, TimeSpanTextBox.ParseFormat, CultureInfo.InvariantCulture);
-            CurrentRecordingStopAfterTimer = GetCurrentStopAfterDispatcherTimeString();
-            TimerForRecordingStopAfterNotRunning = false;
-
-            stopAfterdispatcherTimer.Start();
-        }
-
-        public void ResetStartAfterDispatcherTimer()
-        {
-            startAfterDispatcherTimer.Stop();
-            TimerForRecordingStartAfterNotRunning = true;
-            startAfterDispatcherTimerCurrentTime = TimeSpan.ParseExact(appSettingService.RecordingTimerStartAfterMax, TimeSpanTextBox.ParseFormat, CultureInfo.InvariantCulture);
-            CurrentRecordingStartAfterTimer = appSettingService.RecordingTimerStartAfterMax;
-        }
-
-        public void ResetStopAfterDispatcherTimer()
-        {
-            stopAfterdispatcherTimer.Stop();
-            TimerForRecordingStopAfterNotRunning = true;
-            stopAfterDispatcherTimerCurrentTime = TimeSpan.ParseExact(appSettingService.RecordingTimerStopAfterMax, TimeSpanTextBox.ParseFormat, CultureInfo.InvariantCulture);
-            CurrentRecordingStopAfterTimer = appSettingService.RecordingTimerStopAfterMax;
-        }
-
-        public void ResetAllTimers()
-        {
-            ResetStartAfterDispatcherTimer();
-            ResetStopAfterDispatcherTimer();
         }
 
         public event PropertyChangedEventHandler PropertyChanged;
@@ -525,51 +371,7 @@ namespace RecNForget.WPF.Services
 
         public void ToggleStartStopRecording()
         {
-            // if no timers are configured, reset all timers and toggle record directly
-            if (!appSettingService.RecordingTimerStartAfterIsEnabled && !appSettingService.RecordingTimerStopAfterIsEnabled)
-            {
-                ResetAllTimers();
-                audioRecordingService.ToggleRecording();
-                return;
-            }
-
-            // if currently recording, all timers can be ignored - stop after recording timer is ignored
-            if (audioRecordingService.CurrentlyRecording)
-            {
-                ResetAllTimers();
-                audioRecordingService.ToggleRecording();
-                return;
-            }
-
-            // if NOT currently recording, recording MIGHT be delayed by a timespan set in 
-            if (appSettingService.RecordingTimerStartAfterIsEnabled)
-            {
-                // override start to recording timer if action was triggered again
-                if (startAfterDispatcherTimer.IsEnabled)
-                {
-                    ResetAllTimers();
-                    audioRecordingService.ToggleRecording();
-
-                    if (appSettingService.RecordingTimerStopAfterIsEnabled)
-                    {
-                        StartTimerToStopRecordingAfter();
-                    }
-
-                    return;
-                }
-
-                StartTimerToStartRecordingAfter();
-            }
-            else
-            {
-                ResetAllTimers();
-                audioRecordingService.ToggleRecording();
-
-                if (appSettingService.RecordingTimerStopAfterIsEnabled)
-                {
-                    StartTimerToStopRecordingAfter();
-                }
-            }
+            audioRecordingService.ToggleRecording();
         }
 
         public bool QueueAudioPlayback(string fileName = null, string startIndicatorFileName = null, string endIndicatorFileName = null)
